POF-61: Add tests for custom api endpoints (#11)
* Updated suggestion api's routes to be more logical & fixed one bug

* Added tests for suggestion controller's methods

* Added tests for suggestion/comment endpoint

* Updated tests for suggestion controller

* Improved test structure

* Fixed bug

* Updated grantPriviledge's docs

* Improved tests

+++ utils/algolia.js
@@ -7,6 +7,8 @@ const cleanDeep = require("clean-deep");
  * @param {boolean} draftMode To use draft mode or not (default `true`)
  */
 const updateInAlgolia = (contentType, data, draftMode = true) => {
+  if (process.env.NODE_ENV === "test") return;
+
   const sanitizedData = sanitizeData(data);
 
   if (draftMode) {
@@ -26,6 +28,8 @@ const updateInAlgolia = (contentType, data, draftMode = true) => {
  * @param {string} id Entry's id
  */
 const deleteFromAlgolia = (contentType, id) => {
+  if (process.env.NODE_ENV === "test") return;
+
   strapi.services.algolia.deleteObject(id, contentType);
 };
